# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WedWithMe REST API",
    "version": "1.0.0",
    "description": "Autogenerated OpenAPI specification for the WedWithMe Next.js Backend."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Local API Environment"
    }
  ],
  "paths": {
...
    "/admin/fraud/flags": {
      "get": {
        "summary": "GET /admin/fraud/flags",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/admin/governance/users/{id}": {
      "patch": {
        "summary": "PATCH /admin/governance/users/{id}",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/admin/reports/financial/export": {
      "get": {
        "summary": "GET /admin/reports/financial/export",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/reports/financial/summary": {
      "get": {
        "summary": "GET /admin/reports/financial/summary",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/admin/reviews": {
      "get": {
        "summary": "GET /admin/reviews",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "PUT /admin/reviews",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}